ao_refactor: extract copy button
